Tests/LibWeb/Text/expected/input-time-typing.txt

after '1': value=""
after '2': value=""
after ':': value=""
after '3': value=""
after '0': value="12:30"
